LayoutTests/imported/w3c/web-platform-tests/html/canvas/offscreen/text/2d.text.align.default-expected.txt

2d.text.align.default


FAIL OffscreenCanvas test: 2d.text.align.default assert_equals: ctx.textAlign === 'start' (got [undefined], expected start[string]) expected (string) "start" but got (undefined) undefined
